src/main/java/ru/ssau/tk/kmaster/practice/Tasks/Shipping/Route.java
package ru.ssau.tk.kmaster.practice.Tasks.Shipping;

import java.util.ArrayList;
import java.util.List;

public class Route {

    private final List<Location> locations = new ArrayList<>();

    public List<Location> getLocations() {
        return locations;
    }

    public void addNewLocation(Location newLocation) {
        locations.add(newLocation);
    }

    public void addNewLocation(int index, Location newLocation) {
        locations.add(index, newLocation);
    }

    public void removeLocation(int index) {
        locations.remove(index);
    }
}
